core/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.18
django-filter==24.3
django-mail-templated==2.6.5
django-summernote==0.8.20.0
django-taggit==6.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
